requirements.txt
streamlit
chromadb
sentence-transformers
pypandoc
python-docx
pdfplumber
ollama
tqdm
requests